fix: neither player turn active bug
Under some conditions both players would see "Opponents turn"

This occurred because logic that was supposed to ensure that players
could only end their turn when it was their turn only worked for the
player belonging to the current window and not both players.

So one player would send and end_turn message, it would
work for them, but not for the other player.