ref: Move danger overlay into UI file
Danger overlay represents where AI can attack on the next turn,
rather than drawing it for each unit, we iterate each cell and check if
there are any units who can attack it.  This is simpler to grok,
especially for ranged golems.

created canAttackCell function that returns true if an AI can attack a
cell.