ref: modifiers to support Cards and Events
See card "freeze" for example, it adds a modifier,
the modifier adds an event and the event keeps internal state in
unit.modifiers.freeze until it removes the modifier and itself.

This allows for rich modifiers that use events and have internal state
to do more complicated things like keep track of how many turns to skip
and allow modifiers to stack (freeze and shield are examples of this).