src: big: Make AI actions async function
so that the game can wait for their actions to be complete.

src: Update order of PlayerTurns phase so that players are initialized
at the end of the beginning of the phase.  This is important because
if all players are dead or disconnected, it can immediately end the
player phase, so saving other initialization code to happen after that
can make it all stack up syncronously after a bunch of NPC turns.

src: improvement, move card via turn aquisition into
initializePlayerTurns where it is most suited since it
is to occur at the beginning of player turns.