fix: AI unit image desync when dying while attack
When a unit died while it was attacking (for example if two
units of separate factions moved at the same time) it would
change sprites and stop midway to another cell.  Meaning,
it's image was desynced with it's actual location.  Making
the attack animation "awaitable" (returns a promise rather
than just adding the animation to the timeline) solves this
because now the OTHER AI's attack doesn't deal damage
until the attack animation is finished and since they both attack
at the same time, it will coincidentally wait for all attack movements
to finish before firing Unit.takeDamage