fix: card ui syncing
so that when you get a new card it doesn't clear out
your selected cards.

Also fixes the bug where selected cards would go back
to your hand after casting instead of being removed due to a
race condidtion.  This occurred because only card-hand was cleared,
but if the cards in selected-cards didn't get put back by the time
card-hand was cleared, they'd go back to card-hand after it was
reconciled.