fix: Reset playerTurnIndex only outside
of setTurnPhase.
From endNPCTurnPhase:
Note: The reason this logic happens here instead of in setTurnPhase
is because setTurnPhase needs to be called on game load to put everything
in a good state when updating to the canonical client's game state. (this
happens when one client disconnects and rejoins).  If playerTurnIndex were
reset in setTurnPhase, a client reconnecting would also reset the
playerTurnIndex (which is not desireable because it's trying to LOAD the
game state).  Instead, it's handled here, when the NPC turn phase ends
so that clients CAN reconnect mid player turn and the playerTurnIndex is
maintained