src: Collisions: Handle line and circle collisions
together in the function moveWithCollisions.

Change repelCircle so that the mover doesn't get repelled.  This is
a design decision.  When both circles were repelled equally it resulted
in this odd circular motion behavior because the mover was
continuing to try to get to it's destination frame by frame so
the circles would kind of orbit each other.
Better to just have the mover push others out of the way.