src: desync: fix hash comparison
A hash of an object is different than a hash of a stringified
object.
This ensures that the hash sent via the GAMESTATE_HASH
message can be actually regengerated from hashing the
payload.state.

Add processedMessageCount so that clients can know
ahead of time that their hashes are supposed to be equal,
if the processedMessageCount is not equal, then
one of the clients is lagging behind and will **hopefully**
catch up as it works it's way through the queue.