ref: Setup overhaul
Split LOAD_GAME_STATE into another message so that
first-time clients could load the underworld without waiting
on the queue and existing clients could still
load the gamestate in a queue.

A client being fully ready means that everything is loaded,
the underworld is either initialized (if host client), or loaded
if non-host client, and the player is created (via
SELECT_CHARACTER).  The player must be created AFTER
the underworld is loaded, so I made a second queue just for
SELECT_CHARACTER.