ref: Remove resolveDoneMoving
resolveDoneMoving was a headache in many places and it was only ever
used to end a player's turn after they were done moving, which turned
out to cause a player turn desync in the event that a client joined
mid-move, which meant that they didn't have the promise
that, when resolved, would end the moving players turn.
I'm opting for a quick solution to just loop all players
when a unit is done moving to see if the unit belongs to a player.
This is probably not the best long term solution so I added a TODO.
But for now it resolves the bug and removes the highly-bothersome
"resolveDoneMoving" property of units.