src: add ASK_FOR_INIT_GAME_STATE MESSAGE_TYPE
This commit makes how new clients get the current gamestate
more robust.  Now when a new client joins, the other clients
queue an ASK_FOR_INIT_GAME_STATE into the data queue.
This is important so that they don't get game state that
is mid-way through completing an action. And since all
clients get this in their queue, even if the host
disconnects mid-action, the next client that becomes host
will have it and will send the message when the queue
is ready for it.