ref: Refactor JOIN_GAME
How clients join a game is a hard problem because they need the latest
state and if they join while a message or action is in-flight it can
cause desync problems.  My solution to this is to ensure that a
JOIN_GAME message (send by newly joining clients once they are ready to
join - e.g. they have picked their character and readyState.isReady()
== true) is handled synchronously in the queue along with other onData
messages.  This means that the host will only send the lastest game state
when there aren't any currently running messages.

Also there was complexity in initializing a Player instance for the new client
and syncing the underworld gamestate.  My solution here is to have
the Player isntance initialized in the same message that requests the
latest gamestate.  So the all non-new clients will add the player and
then the host will send the latest gamestate once the new player has been
added.