src: Pathfinding: finish split()
split is pretty complicated, it takes a point an it's connections
(potentially like a hub and spokes) and splits any angles that are
wider than 180degrees by making another connection with
a point (from allPoints) within that angle (e.g. it adds a spoke).
When it is done the angles between spokes are all <= 180 degrees.

It also optimizes the number of connections by removing superfluous
connections if multiple spoke angles can be combined and still not
exceed 180 degrees.