src: Store underworld.pathingWalls
used for generating a walkable path between points.
Pathing walls are like walls but expanded to
account for the thickness of units.  This is so when
units try to move to a point, I can test if their movement
path collides with any of these pathing walls.  If it does
I will divert the movement right or left to the corner
of the wall to see if they can THEN find a straight line path
to their destination.
In this way we can path as needed without generating
a whole convex poly mesh.