fix: Ensure new client only recieves
INIT_GAME_STATE when they are ready.
It goes like this: new client chooses a character and sends a
JOIN_GAME message.  The host recieves the message,
adds the new client as a Player and then sends the whole
game state back as INIT_GAME_STATE.  If the client receives an
INIT_GAME_STATE that does NOT contain a player with their
clientID then they recieved it too early and should disregard it.