fix: Polygon: Floating point error
intersections added as new points should be rounded because otherwise
it can cause pathing to mess up due to floating point.  For example,
I had a scenario where a "straight" vertical line had x values of
27.000000004 and 27.000000007 and so it jumped through a no
walk zone because it could make the line without collision when
it should've been colinear to the wall.