ref: Spell modifiers
Spell modifiers, freeze for example, are used to add state to a unit
when the spell is cast upon them.  So, a unit that has freeze cast on
them will be "frozen" (the modifier).  After manually exporting the
add function so that spells could bee used outside the context of
a card (such as when a vampire bites instead of when YOU cast "bite"
as a vampire), I realized I was manually managing the "add"ing and
"remove"ing of modifiers when it should be done automatically.

For example, when you are un-vampire'd you should lose the "bite"
card.  This happens in the remove function of the modifier.
Since I'm already "register"ing the card, I can also keep track of
the "add" and "remove" functions so all I have to do is invoke
Unit.removeModifier(unit, 'bite'), and it will now handle removing
the subsprites, the modifier state, and also calling the custom
bite remove function which also removes the card.
So now all Spells can export custom add remove functions in
modifier:{add,remove} so that custom logic will execute when the
modifier is added to a unit or removed.