fix: Animation order
Since animateIndependents iterates the animations
in reverse order so that it can remove them as it goes,
it was also executing animations in reverse order
which meant if one animation was called right after the
other, it would technically finish first if they had the same
lerp time.  This was causing a problem where the player unit
was spawning hidden because hide() was called at the end
of the level and show() was called at the beginning (and
in the tutorial we skip the upgrade screen).
This fixes the order that they finish.