ref: mouseMove -> updateMouseUI()
I've found that the mouse ui needs to be updated in more places
than just when the mouse moves, for example, when a hotkey or escape
is pressed, the mouse UI needs to update too.  So i consolidated
all the calls to updates parts of the mouse ui and now they all live
under just updateMouseUI().  Also I check inspectmode on mouse
move so that if the window loses focus while ctrl is depressed, and
the user releases control before the window gets focused again,
it won't get stuck in inspect mode.