src: Add dynamic camera
Camera follows a loose point between the mouse
and the player unit.
Camera scale stays the same size regardless of the level (after
playtesting, we deeteermined it was disorienting to
scale the camera).
- Improve html so that pixiholder's canvas
does underlap the card hand, but this is no longer
an issue since the camera can move now