fix: Prevent queing up multiple endturns
causing upgarde screen to cycle and update while observing it.

I think this might happen because of RESOLVE_DONE_MOVING_TIMEEOUT_MS.
If that triggers the timeout, it can cause the player to end their turn
while they're still moving despite the built in queue that is supposed
to completely process the move and THEN end the turn. If this timeout
was configured properly, the queue would prevent this from being a
problem.
Also I shouldn't have put Upgrade screen initialization logic in setView