fix: projectPointForPathingMesh to consider
the order of prev, point, next.  In this codebase, polygons are
expressed as an array of points. If prev, point, and next go
in a clockwise direction, then the polygon is a normal polygon
where the outside is accessible and the inside is blocked for movement
from units.  However, if the points go in the counter-clockwise direction
the polygon is inverted which means the inside of the polygon shape
is accessible for movement by units and the outside is blocked.

This "inverted" status can be determined by walking the polygon
but it is also hard-coded with the inverted flag.

There was probably a better way to do this in retrospect, but
I'm stuck with the decisions I made then unless I commit to
a huge refactor which I'm not going to do since this works
well enough.

This commit fixes projectPointAlongNormalVector so that
it supports polygons with some concave angles.
Before it was only ever used on perfectly square polygons
(fully convex, with no concave angles); but now that I need
to use it on more complex polygons (the pathing polygons,
which may contain concave angles); I had to fix the algorithm
to support polygons with concave angles.

Now it correctly will find the normal (the outside projection)
of a point regardless if the point has a concave or convex outside
angle