fix: Enemy Attention markers to consider zoom
Add background to card hand so you can't see enemies through it
which would super complicate drawing enemy attention markers
and also might add to user frusteration when they try to click
on an enemy that is behind their health bar.
This is better, it just doesn't allow you to see behind it.