ref: Change upgrade choosing to be handled locally
not over wsPie.  Handling it over wsPie has been causing numerous
desync issues where clients would get stuck.
and view changing is supposed to be local anyway, not synced between
clients.  So this commit changes the view change to be handled
locally, it does not wait for the message to return from wsPie.