fix: Sync player dryRunUnit when health or mana
changes not during a dryRun.  Without these changes
users would temporarily see the diagonal bars in the
health or mana bar because the dryRun unit's health
or mana would still be what it was before the user
picked up the potion or upgrade.
Note: This is a non critical change since
moving the mouse would sync the dryRun units anyway