ref: Underworld setup
Now Players are created as soon as the client joins
the pie room.  This is simpler to reason about
since now they will always have a window.player.

This commit changes it so join happens automatically
and CHANGE_CHARACTER just changes the skin.
It is easier to sync clients when all clients that join
wsPie room have a player.