fix: Bug in polygon merging
that would rarely cause a branch to be undefined
breaking the merge algorithm.
lineSegmentIntersections can be rounded since all
polygon points consist of whole numbers, and
the intersection function sometimes returns
a number like 242.99999999999999991.

In this case, I was rounding the intersection value
in the wrong place, it should've been rounded from
the beginning so that the check if the intersection
was identical to the "otherPolyFirstPoint"
would work correctly