fix: merge issue
Under very specific and rare circumstances, whole groups of obstacles
wouldn't get their polygons into underworld.bounds but strangely they
WOULD get their polygons in to underworld.walls.
I believe this occurred because bounds were reverse expanded
after merging which resulted in odd behavior.  Instead of this,
I am not just creating them directly from obstacle bounds,
not the pathing polygons.

It may be that the previous logic was as it was because I had
faked depth to make walls look tall, but now bounds and walls
directly overlap (unless an obstacle isn't a wall but is a void - not
walkable BUT doesn't block Line of sight).

Anyway, creating the underworld.bounds directly from obstacle bounds
and THEN merging them solves the problem on the following seed:
    this.seed = '0.38655620060114826';