fix: isVec2InsideOfPolygon so that points
on a wall of an inverted polygon count as INSIDE (non-walkable)
the inverted polygon.  This is important in the event that
a non-inverted polygon shares a wall with an inverted polygon.
See the newly added test:
'should merge them without entering an infinite loop'
Without this change to polygon.ts, that test
enters an infinite loop