fixed: Pathing issue
caused by merging polys AFTER expanding polys. I'm not sure why it
happened but it works fine when merging polygons and then
expanding to make the pathing mesh.  I suspect that the
expand then merge caused an issue when processing polygons
that caused them to be marked as "bad polygons"

I used the following to reproduce it:

    this.seed = '0.051284396870881155';
    underworld.initLevel(1)