fix: Race condition with cleanUpLevel
where non-host, if alt tabbed when it went to the next level
would execute cleanUpLevel AFTER createLevel thus removing the pixi
containers and rendering a mostly blank screen.
This change ensures only the createLevel function itself will clean up
the previous level so that this race condition is now impossible.