src: stop throttling runPredictions
It is important to get immediate feedback when queueing up a spell
when that spell will kill an enemy.  I just need to make runPredictions
more efficient so it can run on slower machiens or levels with more
enemies.  Throttling it is not an option - partically because it'll still
cause change when it DOES run (even if running less frequently)