src: Add ONLY_KEEP_LATEST
for preventing queue from holding on to old messages of a same type when
a new message comes through with the latest info.
I found this necessary to add when I had a bunch of PLAYER_THINKING
events queued up while I was a single player in a multiplayer game
and was executing a PLAYER_MOVE.  It would store up like 100
PLAYER_THINKING messages and then just clear process them all
as soon as I was dont moving, but PLAYER_THINKING messages only
matter if they are current, so I could safetly discard the old ones.