src: Prevent MELEE units from moving to
center of target, stop them when they get to the target's edge.
This is done inside of gameLoop instead of inside of
moveTowards because moveTowards used setPath and setPath
(through calculatePath) uses unchanging targets as an optimization
so that target should be the center but this change will still
allow MELEE units to stop at their edge.