fix: big pathing bug.
Melee was stopping short of EVERY point on the path, not just the
final point.  I want them to stop short of the final point so they don't
intersect with their target, since they're trying to get into melee range.
So this fix only applies the "early stop" if the only point in the path is
the last point