fix: circular reference
When the second client joined AFTER the first already picked
their character, the host failed to send the serialized units
due to a circular reference. It turns out, the path.points of
an NPC unit would contain a reference to one of the player
units which made JSON.stringify fail.