fix: the rare archer double damage bug
Since the SET_PHASE message wasn't waiting for
initializeTurnPhase to complete (and since the NPC turn
is executed inside of initializeTurnPhase) it was going and
processing the next message (the unit sync) before
the arrow reached it's target, which synced
the already damaged health over the full health and
THEN the arrow hit on the client2's screen causing it
to appear that the unit took double damage.
This fix awaits initializeTurnPhase so that it won't
process the sync messsage until the Npc turn is fully
complete.
Note: This depends on the NPC's action being awaited
too, so if the archer didn't await the arrow flight
in it's action this could still happen.