fix: processLineSegments: fix infinite loop
When checking to see if the polygon has closed, I have to check that 2
intersections in a row are already in the polygon to ensure it doesn't
stop early when multiple polygons share a single vertex and should
be merged into a single polygon.
I made a mistake previously where it was checking the intersection
and the p2 of the currentLine rather than the intersection and the
next intersection. This commit fixes this.