fix: mergePolygon2s: Test centerpoint instead
of verticies to see if linesegment is enclosed in an other poly.
Because there is a possibility that both verticies could be in a poly
but it's center isn't (see test "...donut of rectancles".  So really to
see if the line segment is fully enclosed, the center point should
be tested