fix: Pathing mesh
The current limitations of mergePolygon2s makes it so I can only
use it on an array of polygons once.  If it use it twice I risk
deleting some of the polygons do to the fact that a taurus polygon
(donut example) isn't aware that it's inside is flipped, so a simple
square intersecting it can take a chunk out that prevents it from
being closed in the processLineSegment step and then it is fully
omitted.
So this work around uses the wall polygons before they are merged
for the first time (so there are no taurus polys) and therefore that issue
does not occur.