ref: Image changing functions
Make them more reliable and convenient. Change sprite now
just takes an imagePath so that I don't have to write
the boilerplate to make a new sprite each time.  It also
won't change the sprite if the sprite is already the right sprite,
which prevents restarting the animation when returnToDefaultSprite
is called. Add imagePath (instead of jid) to ALL sprites to id which
sprite is active.