ref: Image to always have an AnimatedSprite
See Underworld change at line 407 for how to use non animated sprites.
Don't use image, just use builtin Pixi Sprite.

ref: Separate onComplete and resolve
onComplete usually triggers a return to the default sprite and
shouldn't be conflated with resolving whatever is waiting for the animation
to finish because resolve might also need to be invoked
if an animation changes to another animation before it is done
animating, like in the case of a player damaging themself and having
the cast animation change to playerHit.

ref: Update returnToDefaultSprite so that it only changes the sprite
if there isn't a currently playing animation so that it doesn't
clobber back to back non-looping animations.
Also ensure it updates the parent even if the sprite doesn't change
because it is sometimes used to ensure the parent is the correct sprite

src: Image: Always trigger resolver when sprite.onComplete occurs so
that I don't have to add it in two places when invoking changeSprite