fix: returnToDefaultSprite
Sometimes returnToDefaultSprite does need to override a looping
animation, like when a walk animation needs to change to either a
death animation OR the default idle animation.

This check (!unit.image.sprite.playing) should never have been added
to returnToDefaultSprite. Instead, returnToDefaultSprite should've been
called within an animation's onComplete (which, in the case of one animation
being replaced with another animation will prevent idle from
clobbering the second animation because the first never finished and
never calls onComplete - but now that resolve is independent of
onComplete it won't hang, waiting for the promise to resolve. This is much better.
The animation promise resolve is decoupled from onComplete which is
reserved ONLY for when a non-looping animation reaches it's end.
Whereas resolve needs to be called both when an animation completes
(which is now done automatically in changeSprite AND when an animation
is replaced - in which case onComplete SHOULDN'T be invoked.