ref: Improve outOfRange handling
replace all_red shader with pixi's tint.
outOfRange spells are now handled for maximum player convenience.
If there is a viable target at the end of the player's range, even if
their mouse is out of range it will try to predict a cast (and allow
a cast) at the end of their range.  If there are no targets at end of
range, it will  show a prediction for out of range and show
targets with grey tint to suggest they are out of range.