art: Scale down gamplay and restore unit original scale
I originally scaled units up because of the "scale of the gameplay"
(ranges and how far things could move).  But the art was
designed to be a specific size, especially relative to the tiles
so this commit returns the scale the normal
and scales down the gameplay along with it.