fix: If path cannot be made, path to nearest point
on wall.  This occurs when, for example, a player is trying
to path to a point that is inaccessible because it is through
a pathing wall.
Easy way to reproduce: Get right on a pathingLineSegment(pink lines)
and try to path to the brick portion of a liquidSGroundN.png tile
and you'll see it draws a straight line rather than a path and the
unit does not move