fix: Errant early resolve
This extra protection isn't needed. resolveDoneMoving is only set
in this function and it is covered by raceTimeout.
Also this make it so that the walk animation never played, because
for player movment, moveTowards is called multiple times and
so it would resolve the promise that triggered the walk animation
and set it back to default sprite (idle)