fix: corner case where liquid interactions didn't work
This occurred when the distance pushed was just so that
it traveled far enough so that the units original position was out of
the liquid and the moved position was inside the threshold so that it would
think they were already in the liquid and pop them back out.

Now it considers a line from their last position to their current
position, so no matter how fast or far they are moving, it will
detect if they pass through liquid bounds.
Also optimized fn invokation so it is only invoked when a unit is moving
of it's own path (so that they can climb out of liquid) or when they are
force pushed.  No longer does it run for every unit on every gameloop.