ref: Satisfied TS for headless
I've been really struggling to get both sets of tsconfigs
working on the same codebase.  It's unfortunate that there's
not a better solution but setting all of the globals that are in
the client version but not in headless to "| undefined" is the only
solution I've found that works.
I was hoping I could just leave them as strictly undefined but then
typescript gives me "never" errors that I can't get past.