ref: Remove readyState
Setup is handled more orderly now.
For browser clients, setupAll sets up PIXI first (because its async
and everything depends on it), when its done, it sets up
underworld and Pie. Underworld constructor sets up content (cards and
units). And messages will only be sent if a room is joined anyway
so wsPieConnection and wsPieRoomJoined should be fine.

Now there is a single underworld for any instance of the app running
and it is never cleaned up so readyStates' pixiAssets, underworld, and
 content will always be ready after the game is setup the first time.

For headless server, the host app is made inside of makeHostAppInstance
which implicitly makes a pie room so wsPieConnection and wsPieRoomJoined
are a given. Instantiating underworld still registers content and it
doesn't need pixi so it's ready state is just the underworld, content, and
pie connections.