src: Add inLiquid modifier to manage
the inLiquid state (if a unit has the inLiquid mask).
Optimization: tryFallOutOfLava checks to see that
the unit has the inLiquid modifier before checking
so it doesn't waste cpu on checking on inLiquid units
to see if they've left the liquid bounds
src: Add prediction arg to modifier's add().
This is needed for inLiquid so it can pass along prediction
to takeDamage