fix: bloodcurse causing wrong health upgrade text
Since blood curse is a modifier it is removed at the end of a
level. And at the end of a level you may get a health upgrade
offered. The modifiers must be removed BEFORE the upgrades
are generated or else it will tell you that you'll get the wrong
health