ref: Simplify "coyote range"
Allow clicks near end of cast range to be considered
a click right at end of range.
This is for player convenience so they don't have to be
pixel perfect when trying to cast on a target at their end
range.
It also removes a bunch of overly complicated code
that tried to manage casting at end of range
if there was a target there while your mouse was beyond
it.  This is too much, now it just changes the entire cast line to
grey when it's out of range.