fix: pathfinding so it always returns the next destination
as the first point.

In the condition that it had to just return a path to the first
intersection it was also returning the start point as the first point
which is superfluous since the unit pathing is already at the start
point. This commit fixes that in this specific condition by returning
a path with only the end point since it's a straight line