fix: headless server crash
due to predictionGraphics being not defined.
"ReferenceError: predictionGraphics is not defined"

Referencing the full "globalThis.predictionGraphics" fixes the crash
AND setting it to undefined in Shims.ts prevents the crash
if it's just referenced as "predictionGraphics"
I believe this is what caused the mysterious network disconnect
during my multiplayer playtest the other day.