ref: Connect so that is is entity-ambiguous
Add new functions: getPotentialTargets() to return an array of any
entities that can act as targets (and it handles prediction logic)
and getCurrentTargets() which gets the targets of the current spell.

These can now be used to return ambiguous entities which can
later be cast into Unit or Pickup or etc (doodad is upcoming) so that
spell logic (such as Connect) can be written to work for all entities that
share common properties like (targetable or HasSpace) this reduces
the surface area for bugs because now theres only a few small places
that the specific logic needs to be added to have a new type of entity
that also responds to these spells.

So for example when I add Doodads (the rock, tree, urn, etc); they are
neither pickups nor units but they share some properties like HasSpace.
So all I have to do is change those two functions and add an isDoodad
typeguard function and add doodadPrediction array
and then chain will work out of the box;