fix: bug where player units persist between
playthroughs in the units array.
This occurred because the units where never cleaned up
in underrworld.cleanup and AI units are automatically
cleaned up when underworld.cleanUpLevel is called
and so the player unit was left in the units array but
without an image and they could be resurrected which
restored the image.