fix: Headless issue with requestAnimationFrame
Headless server should execute all of it's loops as fast
as possible.  It should NOT use requestAnimationFrame.
For example, fly() was causing a
`RangeError: Maximum call stack size exceeded`
because it was invoking requestAnimationFrame with no delay which
caused a ton of recursion. Fly  uses requestAnimationFrame to animate
the projectile but on the headless server this should happen
immediately. So I adapted fly() and lob() to finish immediately on
headless and put a warning in requestAnimationFrame to report
if it is ever invoke from headless again that it needs to be optimized
to execute immediately