fix: forceMove predictions
forceMove predictions were broken when a second order push
occurred.  There were a few things causing this problem:
1. Clone didn't assign the clone a new id on prediction because
prediction doesn't have it's own lastUnitId.  So any logic that depended
on 2 units not sharing the same id was inconsistent prediction vs
reality.
2. forceMoves were simulated all at once as soon as they were added to
the array so if you had 2 targets and they each received a push, it would
fully execute the first, then fully execute the 2nd rather than
executing them together after they are added like it is done in reality.
I fixed this by moving fullySimulateForceMovePredictions() out
of where the forceMoves were made and into castCards().
3. Clone was broken with prediction because it depends on
getting a reference back from Unit.load; however, in prediction,
Unit.load calls addUnitToArray which adds a copy of the unit
passed in into this.unitsPrediction and so the unit that Unit.Load
returned was actually a reference to a different object than the one
that is added to unitsPrediction.