ref: Rework curses to have less room for implementation errors with quantity.
Utility function getOrInitModifier handles quantity both on first
invokation of the curse or blessing and on future invokations.
Spells like shield that request special properties that change
independent of the quantity of the spell applied can still add
their own properties to the modifier object.

This change abstracts the commonly required modifier object
initialization and mutation so that there is less room for
implementation errors.