fix: portal not showing up for players
that join after it has spawned.
This was due to headless server NOT handling scroll cleanup - where
it flies to each player to give them the pickup and then spawns
the portal.
I believe the headless server was just waiting indefinitely due to not
having a globalThis.player.
Added a raceTimeout to catch this kind of thing in the future,
but based on the logic I just added 'spawnPortalFlyScrolls' should NEVER timeout