fix: GC memory leak for PIXI Text
PIXI.Text needs to be destroyed manually or else it
will hold on to a reference to the texture.
https://www.html5gamedevs.com/topic/31749-how-to-cleaning-up-all-pixi-sprites-and-textures/?do=findComment&comment=182386

To reproduce, run
`setInterval(() => {console.log(Object.keys(globalThis.pixi.utils.BaseTextureCache).length);},50);`
and then create text instances.

Note: Other textures do NOT need to be destroyed since so long as a user
is running the game, they need all the sprite textures loaded; whereas,
text sprites are one-off textures and need to be cleaned up after use.