ref: make overworld.underworld optional
which better reflects the reality that there are times when there
should be no underworld that can be acted on.  Such as when the
player is in the main menu, when they quit a game, etc.
This way console.error will report if an action is attempted
while the underworld doesn't exist when it should exist.

This also solves the problem that has existed since the overworld
was first introduced where quitting a game wouldn't take
the player back to the main menu because the cleanup logic
was messed up due to the recursive nature of the underworld
referencing the overworld and the overworld ALWAYS needing
a valid underworld object.