fix: this.endFullTurnCycle wasn't await'd
This caused an issue with onEndTurn events.
For example, I'm adding summoner sickness and somehow
4 of the 5 units summoned would properly have their onTurnEnd
summoner sickness events called but it seemed like the first one of
the 5 didn't.  It turns out, it did, but it was executed and then
broadcastTurnPhase was invoked which caused a unit sync which overwrote
the state of that unit.
This fix properly awaits the endTurnEvents before broadcasting the next
turn phase