chore: Support storage for electron app
This is pretty complicated due to how electron
communicates between the main process and the
renderer. but basically electron pulls the stored file
and exposes it's contents in a global variable.
Then the renderer makes a copy of that global
variable (because it's immutable) and just updates
the copy whenever anything stored changes.
(It also sends the stored changes to electron to
persist to disk, but that is an async operation).

So the copy allows us to have syncronous access to stored information.