chore: Support loading settings from disk
for an electron build.
After many hours of trying different ways, I cannot
get electron to load settings from disk and give it
to the front end syncronously without enabling
nodeIntegration which i do not want to do for security
reasons so async it is.  This means that when the
settings first load they won't be from disk so
once the settings do load from disk it will need
to override settings such as volume and language.