fix: missing wall pathing polygons
that occurred as a regression due to the new
stamping liquid method.
When a liquid pool is stamped right up against the border
of the map it didn't make walls along all the edges
which resulted in a "hole" which made the wall
pathing polygons fail to generate under those conditions.