fix: infinite loop when NPC Ally is last ally alive.
By simplifying the places where NPC_ALLY turn can start from 2 places
to 1, I resolve the issue that occured in the following scenario:

You die and you have an ally npc.  The game continues, the ally npc wins.
The turn_phase enters an infinite loop.

Now when you die it will still end your turn. Ending your turn checks
if it should go to the next level, but the NPC is still alive so it
tryEndPlayerTurnPhase() and does going to NPC_ALLY phase.
When the npc ally kills the remaining NPC_ENEMY, it goes to PlayerTurns
which ends your turn again which checks for end of level again
and this time it does go to the end of level.

There's a lot of complexity here, but ensuring that there is only ONE PLACE
that each turn_phase can be broadcast from ensures that there are
no special code paths that skip checks.

Note: PlayerTurns is the exception, it may be called from different
locations.
1. in postSetupLevel to start the turn phase cycling
(this is an entrypoint)
2. in tryRestartTurnPhaseLoop().  This is also an entry point for when
the server exits the Stalled phase
3. AFter the NPC_ENEMY turn, this is the natural way that it cycles back to
PlayerTurns