fix: targeting spell + move spell causing desync
Since each spell has to fully complete before the next spell,
triggerGameLoopHeadless should occur immediately after each
card.effect so that the headless server can fully simulate any
forceMoves.  If it doesn't, it will racetimeout the forceMove
because, in the previous code, it was syncronously calling handlespell
expecting the forcemoves to be added syncronously, but
because they were in the second spell (in chain + repel),
it continued execution, hit triggerGameLoopHeadless and there
were nothing to simulate so it awaited the spell promise and THEN
it added the forceMoves and the await timed out and they never moved.

To reproduce, go to the old commit, cast connect+repel in multiplayer and
end your turn and you'll witness desync.

This commit fixes the issue by correctly and fully simulating each card
after its effect is invoked.