ref: Support card cost refunds for fizzle spells
Totally refactor spell cost so it calculates one at a time.
If a spell fizzles (which must be manually detected in each
spell effect fn) it will refund the mana or health cost and
not increment the card usage.

clean: Remove updateManaCostUI.  The cost calculation should just be
moved outside it since it was only used in one instance and then it was just
a oneliner that made sense to be inlined.