ref: maintain unit prediction object ref
after calling syncPlayerPredictionUnitOnly so that
effectState.targetedUnits will contain the same ref
to the prediction unit as was established the last time
syncPredictionEntities was called.

This change was made due to the previous commit's
solved issue of trying to figure out why a blue death skull
didn't appear over the players head when the player hovered
a deadly spell above themself - which led me to discover
that the player's unitPrediction was no longer the same
as the unitPrediction stored in effectState.targetedUnits
when predicting the outcome of a spell.