ref: Make casters immovable while casting
This prevents multiple prediction upset bugs where
if you cast and then move before the cast is complete
the result will be different from the prediction.

UX: Add casting particles coming off of your character while casting
to denote that a cast is ocurring.
ref: Move containerParticlesUnderUnits to Particles.ts
and actually make it a ParticleContainer.